Refactoring utilities
